refactor: /src/components/pages/profile 폴더의 구조 개선
