+++ src/components/pages/profile/index.ts
@@ -1,2 +1,3 @@
 export * from './Header';
 export * from './ProfileBasicInformationForm';
+export * from './Stepper';
